+++ User-Stories.excalidraw
@@ -220,6 +220,34 @@
       "locked": false
     },
     {
+      "type": "rectangle",
+      "version": 1469,
+      "versionNonce": 569494213,
+      "isDeleted": false,
+      "id": "ex87y2z6WfWrbKAEYjrOC",
+      "fillStyle": "hachure",
+      "strokeWidth": 1,
+      "strokeStyle": "solid",
+      "roughness": 1,
+      "opacity": 100,
+      "angle": 0,
+      "x": 213.81941819190354,
+      "y": 2041.8893606894628,
+      "strokeColor": "#000000",
+      "backgroundColor": "transparent",
+      "width": 646.6776060677006,
+      "height": 869.2936672251318,
+      "seed": 68218539,
+      "groupIds": [],
+      "roundness": {
+        "type": 3
+      },
+      "boundElements": null,
+      "updated": 1679323114487,
+      "link": null,
+      "locked": false
+    },
+    {
       "type": "text",
       "version": 1564,
       "versionNonce": 1864237115,
@@ -425,6 +453,40 @@
     },
     {
       "type": "text",
+      "version": 6928,
+      "versionNonce": 956220997,
+      "isDeleted": false,
+      "id": "WThv70T7SJkWMqEQ9V1Ep",
+      "fillStyle": "hachure",
+      "strokeWidth": 1,
+      "strokeStyle": "solid",
+      "roughness": 1,
+      "opacity": 100,
+      "angle": 0,
+      "x": 256.99571134366136,
+      "y": 2105.661147282118,
+      "strokeColor": "#000000",
+      "backgroundColor": "transparent",
+      "width": 566,
+      "height": 775,
+      "seed": 884663685,
+      "groupIds": [],
+      "roundness": null,
+      "boundElements": null,
+      "updated": 1679323219952,
+      "link": null,
+      "locked": false,
+      "fontSize": 20.78505821606629,
+      "fontFamily": 1,
+      "text": "## User Story 7\n\n## Starting Page\n\nAs an User\nI want a start screen  \nSo that the data is loaded when i got to the \nhomepage \n\n## Description\n\n- new Page\n- after 3 seconds the user is pushed to homepage\n\n## Acceptance criteria\n\n- [ ] new page\n- [ ] first page the user is seeing \n- [ ] after 3 seconds he gets redirected to Homepage\n- [ ] a good styling\n\n## Tasks\n\n- [ ] add new page\n- [ ] change the routes\n- [ ] use a lottiefile to add something nice\n- [ ] style the loading div \n\n\n",
+      "baseline": 768,
+      "textAlign": "left",
+      "verticalAlign": "top",
+      "containerId": null,
+      "originalText": "## User Story 7\n\n## Starting Page\n\nAs an User\nI want a start screen  \nSo that the data is loaded when i got to the \nhomepage \n\n## Description\n\n- new Page\n- after 3 seconds the user is pushed to homepage\n\n## Acceptance criteria\n\n- [ ] new page\n- [ ] first page the user is seeing \n- [ ] after 3 seconds he gets redirected to Homepage\n- [ ] a good styling\n\n## Tasks\n\n- [ ] add new page\n- [ ] change the routes\n- [ ] use a lottiefile to add something nice\n- [ ] style the loading div \n\n\n"
+    },
+    {
+      "type": "text",
       "version": 218,
       "versionNonce": 1355067962,
       "isDeleted": false,
@@ -465,34 +527,6 @@
       "verticalAlign": "top",
       "containerId": null,
       "originalText": "User Stories"
-    },
-    {
-      "type": "rectangle",
-      "version": 379,
-      "versionNonce": 281159270,
-      "isDeleted": false,
-      "id": "sxRW3cppYur-MeuoB2Z0k",
-      "fillStyle": "hachure",
-      "strokeWidth": 1,
-      "strokeStyle": "solid",
-      "roughness": 1,
-      "opacity": 100,
-      "angle": 0,
-      "x": 128.68113760015558,
-      "y": -22.689133967674024,
-      "strokeColor": "#000000",
-      "backgroundColor": "transparent",
-      "width": 2120.1284657384713,
-      "height": 2429.669262272004,
-      "seed": 1622922785,
-      "groupIds": [],
-      "roundness": {
-        "type": 3
-      },
-      "boundElements": [],
-      "updated": 1678649804874,
-      "link": null,
-      "locked": false
     },
     {
       "type": "arrow",
